--- a/torch/__init__.py
+++ b/torch/__init__.py
@@ -552,6 +552,7 @@ import torch.nn
 import torch.nn.intrinsic
 import torch.nn.quantized
 import torch.optim
+import torch.optim._multi_tensor
 import torch.multiprocessing
 import torch.sparse
 import torch.utils.backcompat