--- a/torch/optim/_multi_tensor/__init__.pyi
+++ b/torch/optim/_multi_tensor/__init__.pyi
@@ -1,0 +1,8 @@
+from .adam import Adam as Adam
+from .adamw import AdamW as AdamW
+from .sgd import SGD as SGD
+from .rmsprop import RMSprop as RMSprop
+from .rprop import Rprop as Rprop
+from .asgd import ASGD as ASGD
+from .adamax import Adamax as Adamax
+from .adadelta import Adadelta as Adadelta