--- a/torch/optim/_multi_tensor/rprop.pyi
+++ b/torch/optim/_multi_tensor/rprop.pyi
@@ -1,0 +1,5 @@
+from typing import Tuple
+from ..optimizer import _params_t, Optimizer
+
+class Rprop(Optimizer):
+    def __init__(self, params: _params_t, lr: float=..., etas: Tuple[float, float]=..., step_sizes: Tuple[float, float]=...) -> None: ...