--- a/torch/optim/_multi_tensor/sgd.pyi
+++ b/torch/optim/_multi_tensor/sgd.pyi
@@ -1,0 +1,4 @@
+from ..optimizer import _params_t, Optimizer
+
+class SGD(Optimizer):
+    def __init__(self, params: _params_t, lr: float, momentum: float=..., dampening: float=..., weight_decay:float=..., nesterov:bool=...) -> None: ...