… test/test_foreach.py
… test/test_optim.py
… torch/optim/_multi_tensor/rprop.py