@izdeby 2020-09-28
> > Original PR #45299. The present PR fixes minor bugs that caused revert.
> 
> What's the patch you applied in the end? Is there a test to ensure this is not happening?

All that was tested [via](https://github.com/pytorch/pytorch/pull/45408/commits/3982f300aaaf1a50c59336c9f33084b025512285)

Answering your question, i have additionally tested each optimizers by running https://github.com/pytorch/vision/blob/master/references/classification/train.py  these params: `clear && python references/classification/train.py -j 2 -b 3 --epochs 2 --device cuda `, which resulted in this [update](https://github.com/pytorch/pytorch/pull/45408/commits/cfb3147e7c77fd3f89164e2c4bac08f05c807b9d).