@ngimel 2020-09-28
My suggestion for benchmarking is comparing mta optimizers with regular optimizers operating on a single large parameter (that results from catting all the network parameters). Properly implemented mta optimizers should be pretty close to that, slower, but not by much. Such benchmarks can alert us to lingering loops in the mta implementation.