@vincentqb 2020-09-28
> My suggestion for benchmarking is comparing mta optimizers with regular optimizers operating on a single large parameter (that results from catting all the network parameters). Properly implemented mta optimizers should be pretty close to that, slower, but not by much. Such benchmarks can alert us to lingering loops in the mta implementation.

oops, did I misunderstand the context of the numbers mentioned in the description above? i was under the impression this is what they covered. maybe you meant with even larger model? @izdeby @ngimel