@rgommers 2020-10-02
Hi @izdeby, I noticed this PR introduced a new set of `.pyi` stub files. This was undesired, we're almost done getting rid of stub files and inlining type annotations in `.py` files. See https://github.com/pytorch/pytorch/wiki/Guide-for-adding-type-annotations-to-PyTorch. The `torch.optim` ones are a few of the last ones left, and they're about to be removed in gh-43089.

Would you mind moving these new annotations inline?